only create repl set and add it as a shard when newly created

+++ virt/docker/start_shard_cluster.sh
@@ -9,7 +9,9 @@ docker-port() {
   docker port $@ 27017|cut -d ":" -f2
 }
 
-SHARDS=${1:-2}
+DEFAULT="2"
+read -e -i "$DEFAULT" -p "How many shards do you want to create : " INPUT
+SHARDS="${INPUT:-$DEFAULT}"
 
 echo "Creating Mongodb cluster with $SHARDS shards."
 
@@ -200,16 +202,20 @@ do
     echo "Container rs${i}_srv3 exists."
   fi
 
-  echo "Creating rs${i}"
-  rm -f rs${i}.js
-cat <<EOF > rs${i}.js
-  config = {_id: 'rs${i}', members: [
-  {_id: 0, host: '$(docker-ip rs${i}_srv1)'},
-  {_id: 1, host: '$(docker-ip rs${i}_srv2)'},
-  {_id: 2, host: '$(docker-ip rs${i}_srv3)'}]
-}
-rs.initiate(config);
-EOF
+  ID=$(mongo --port $(docker-port rs${i}_srv1) --eval "printjson(rs.status().ok)" | tail -1)
+  if [ "$ID" -eq "1" ]; then
+    echo "Replication set rs${i} is active."
+  else
+    echo "Creating replication set rs${i}"
+    rm -f rs${i}.js
+    cat <<EOF > rs${i}.js
+    config = {_id: 'rs${i}', members: [
+    {_id: 0, host: '$(docker-ip rs${i}_srv1)'},
+    {_id: 1, host: '$(docker-ip rs${i}_srv2)'},
+    {_id: 2, host: '$(docker-ip rs${i}_srv3)'}]
+  }
+  rs.initiate(config);
+  EOF
   mongo --port $(docker-port rs${i}_srv1) < rs${i}.js
   attempt=0
   while [ $attempt -le 59 ]; do
@@ -234,13 +240,19 @@ EOF
     fi
     sleep 2
   done
+fi
 
+ID=$(mongo --port $(docker-port mongos1) --eval "printjson(sh.status())")
+if grep -q "rs${i}" <<< $ID ; then
+  echo "Shard rs${i} is present."
+else
   echo "Adding rs${i} as shard to cluster."
   rm -f sh.js
-cat <<EOF > rs${i}_sh.js
-sh.addShard("rs${i}/$(docker-ip rs${i}_srv1):27017")
-EOF
+  cat <<EOF > rs${i}_sh.js
+  sh.addShard("rs${i}/$(docker-ip rs${i}_srv1):27017")
+  EOF
   mongo --port $(docker-port mongos1) < rs${i}_sh.js
+fi
 done
 
 echo "MongoDB Cluster is now ready to use."
